scripts/requirements.txt
Pillow>=11.1.0
requests>=2.25.0